sr(api): move get-training in devcomp
