+++ api/tests/devcomp/unit/domain/usecases/get-training_test.js
@@ -1,6 +1,6 @@
-import { expect, sinon, catchErr } from '../../../test-helper.js';
-import { getTraining } from '../../../../lib/domain/usecases/get-training.js';
-import { NotFoundError } from '../../../../lib/domain/errors.js';
+import { expect, sinon, catchErr } from '../../../../test-helper.js';
+import { getTraining } from '../../../../../src/devcomp/domain/usecases/get-training.js';
+import { NotFoundError } from '../../../../../lib/domain/errors.js';
 
 describe('Unit | UseCase | get-training', function () {
   let trainingRepository;
